Fixed an issue not recalculating key count in each node in case at least a node splits. Both gtest and integration test under tests path passed. Test done in both Release and Debug mode.

+++ src/nldb/nldb_txevent.h
@@ -212,7 +212,8 @@ class TxTransactionLogPublisher : public disruptor::EventHandlerInterface<TxTran
 
 		if ((sequence & 0xFFFFF) == 0 )
 		{
-			printf("Log publisher (sequence = %lld): published replication messages to slaves.\n", sequence);
+			printf("Log publisher (sequence = %lld): published replication messages to slaves.\n",
+					(unsigned long long int)sequence);
 		}
 
 		event->getLogBuffer().destroyReplicationMessage(&replicationMessage);
